+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Pink palettes and swatches - [#53](https://github.com/dartoos-dev/eo_color/issues/53)
-  - **Palettes:** Pink, PinkAccent.
-  - **Swatches:** Pinks, PinkAccents.
+- Red, Pink palettes and swatches - [#57](https://github.com/dartoos-dev/eo_color/issues/57)
+  - **Palettes:** Pink, PinkAccent; Red, RedAccent.
+  - **Swatches:** Pinks, PinkAccents; Reds, RedAccents.
 
 ## [0.0.9] - 2021-05-21
 ### Added